chore(ui): consolidate Loading indicator
